requirements.txt
asgiref                        
black                          
click                           
colorama                      
Django                        
django-cors-headers           
djangorestframework           
djangorestframework-simplejwt 
flake8                        
mccabe                       
mypy-extensions              
pathspec                    
pillow                      
pip                         
platformdirs               
psycopg2-binary                
pycodestyle                 
pyflakes                    
PyJWT                     
pytz                        
setuptools                    
sqlparse                      
tomli                      
tzdata    
wheel